Performed refactoring on React Tailwind scaffolding pages
