+++ lib/pages/tailwind/profile.js
@@ -1,0 +1,241 @@
+function profileSource(useTS) {
+  return `import { useState } from 'react';
+  export default function Profile() {
+    const [isEditing, setIsEditing] = useState(false);
+    const [profile] = useState({
+      name: 'Nirmal Samaranayaka',
+      email: 'nirmal.fullstack@gmail.com',
+      phone: '+46 (72) xxx-xxxx',
+      location: 'Stockholm, Sweden',
+      company: 'Scania AB.',
+      position: 'Senior Fullstack Developer',
+      education: 'Computer Science, University of Colombo',
+      website: 'https://dev.to/nirmalsamaranayaka',
+      bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.'
+    });
+    const [settings, setSettings] = useState({
+      emailNotifications: true,
+      pushNotifications: false,
+      darkMode: false,
+      twoFactorAuth: true
+    });
+    const [skills] = useState([
+      'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends', 
+      'TypeScript', 'JavaScript', 'HTML/CSS', 'jQuery', 'Tailwind CSS', 'Bootstrap',
+      '.NET 6/7/8', 'ASP.NET Core', 'C#', 'Web API', 'Microservices', 'Azure', 'AWS'
+    ]);
+    const handleSave = () => {
+      setIsEditing(false);
+    };
+    const handleCancel = () => {
+      setIsEditing(false);
+    };
+    const handleSettingChange = (setting ${useTS?': string':''}) => (event ${useTS?': React.ChangeEvent<HTMLInputElement>':''}) => {
+      setSettings(prev => ({
+        ...prev,
+        [setting]: event.target.checked
+      }));
+    };
+    return (
+      <div className="min-h-screen bg-gray-50">
+        <div className="container mx-auto px-4 py-8">
+          <div className="flex justify-between items-center mb-8">
+            <h1 className="text-4xl font-bold text-gray-900">Profile</h1>
+            <div className="flex space-x-3">
+              {isEditing ? (
+                <>
+                  <button 
+                    className="bg-blue-600 hover:bg-blue-700 text-white px-6 py-3 rounded-lg transition-colors"
+                    onClick={handleSave}
+                  >
+                    Save
+                  </button>
+                  <button 
+                    className="bg-gray-300 hover:bg-gray-400 text-gray-700 px-6 py-3 rounded-lg transition-colors"
+                    onClick={handleCancel}
+                  >
+                    Cancel
+                  </button>
+                </>
+              ) : (
+                <button 
+                  className="bg-blue-600 hover:bg-blue-700 text-white px-6 py-3 rounded-lg transition-colors"
+                  onClick={() => setIsEditing(true)}
+                >
+                  Edit Profile
+                </button>
+              )}
+            </div>
+          </div>
+          <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
+            {/* Profile Header */}
+            <div className="lg:col-span-3">
+              <div className="bg-white rounded-lg shadow-lg p-8">
+                <div className="flex items-center space-x-8">
+                  <div className="w-32 h-32 bg-blue-600 text-white rounded-full flex items-center justify-center text-4xl font-bold">
+                    {profile.name.split(' ').map(n => n[0]).join('')}
+                  </div>
+                  <div className="flex-1">
+                    <h2 className="text-3xl font-bold text-gray-900 mb-3">{profile.name}</h2>
+                    <h4 className="text-xl text-gray-600 mb-4">{profile.position} at {profile.company}</h4>
+                    <p className="text-gray-600 mb-6">{profile.bio}</p>
+                    <div className="flex flex-wrap gap-2">
+                      {skills.map((skill) => (
+                        <span key={skill} className="bg-gray-100 text-gray-700 px-3 py-1 rounded-full text-sm">
+                          {skill}
+                        </span>
+                      ))}
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+            {/* Profile Details */}
+            <div className="lg:col-span-2">
+              <div className="bg-white rounded-lg shadow-lg">
+                <div className="p-6 border-b border-gray-200">
+                  <h3 className="text-xl font-semibold text-gray-900">Personal Information</h3>
+                </div>
+                <div className="p-6">
+                  <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Full Name</label>
+                      <input
+                        type="text"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.name}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Email</label>
+                      <input
+                        type="email"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.email}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Phone</label>
+                      <input
+                        type="tel"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.phone}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div>
+                      <label className="text-sm font-medium text-gray-700 mb-2">Location</label>
+                      <input
+                        type="text"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.location}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Company</label>
+                      <input
+                        type="text"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.company}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Position</label>
+                      <input
+                        type="text"
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        value={profile.position}
+                        disabled={!isEditing}
+                      />
+                    </div>
+                    <div className="md:col-span-2">
+                      <label className="block text-sm font-medium text-gray-700 mb-2">Bio</label>
+                      <textarea
+                        className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        rows={3}
+                        value={profile.bio}
+                        disabled={!isEditing}
+                      ></textarea>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+            {/* Settings */}
+            <div className="lg:col-span-1">
+              <div className="bg-white rounded-lg shadow-lg">
+                <div className="p-6 border-b border-gray-200">
+                  <h3 className="text-xl font-semibold text-gray-900">Settings</h3>
+                </div>
+                <div className="p-6">
+                  <div className="space-y-4">
+                    <div className="flex items-center justify-between">
+                      <span className="text-sm font-medium text-gray-700">Email Notifications</span>
+                      <input
+                        type="checkbox"
+                        className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
+                        checked={settings.emailNotifications}
+                        onChange={handleSettingChange('emailNotifications')}
+                      />
+                    </div>
+                    <div className="flex items-center justify-between">
+                      <span className="text-sm font-medium text-gray-700">Push Notifications</span>
+                      <input
+                        type="checkbox"
+                        className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
+                        checked={settings.pushNotifications}
+                        onChange={handleSettingChange('pushNotifications')}
+                      />
+                    </div>
+                    <div className="flex items-center justify-between">
+                      <span className="text-sm font-medium text-gray-700">Dark Mode</span>
+                      <input
+                        type="checkbox"
+                        className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
+                        checked={settings.darkMode}
+                        onChange={handleSettingChange('darkMode')}
+                      />
+                    </div>
+                    <div className="flex items-center justify-between">
+                      <span className="text-sm font-medium text-gray-700">Two-Factor Auth</span>
+                      <input
+                        type="checkbox"
+                        className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
+                        checked={settings.twoFactorAuth}
+                        onChange={handleSettingChange('twoFactorAuth')}
+                      />
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div className="bg-white rounded-lg shadow-lg mt-6">
+                <div className="p-6 border-b border-gray-200">
+                  <h3 className="text-xl font-semibold text-gray-900">Quick Actions</h3>
+                </div>
+                <div className="p-6">
+                  <div className="space-y-3">
+                    <button className="w-full bg-gray-100 hover:bg-gray-200 text-gray-700 px-4 py-2 rounded-lg transition-colors">
+                      Download Resume
+                    </button>
+                    <button className="w-full bg-gray-100 hover:bg-gray-200 text-gray-700 px-4 py-2 rounded-lg transition-colors">
+                      Share Profile
+                    </button>
+                    <button className="w-full bg-red-100 hover:bg-red-200 text-red-700 px-4 py-2 rounded-lg transition-colors">
+                      Delete Account
+                    </button>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+  
+   );
+  }`;
+  }
+ module.exports = { profileSource };
