+++ lib/scaffold-react-project-structure.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 const { createMUIPages } = require("./pages/mui");
 const { createBootstrapPages } = require("./pages/bootstrap");
+const { createTailwindPages } = require("./pages/tailwind");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -140,580 +141,6 @@ export const router = createBrowserRouter([
     // Default to MUI if no specific UI is chosen
     createMUIPages(pagesDir, extx, ext, useTS);
   }
-}
-
-function createTailwindPages(pagesDir, extx, ext, useTS) {
-  const home = path.join(pagesDir, `Home.${extx}`);
-  if (!fs.existsSync(home)) {
-    writeFileSafe(home, `export default function Home() {
-  const features = [
-    { icon: '🚀', title: 'Performance', description: 'Lightning-fast builds with modern tooling', color: 'text-blue-600' },
-    { icon: '👥', title: 'Team Ready', description: 'Built for collaboration and scalability', color: 'text-green-600' },
-    { icon: '⚡', title: 'Fast Development', description: 'Hot reload and instant feedback', color: 'text-purple-600' },
-    { icon: '🛡️', title: 'Type Safe', description: 'Full TypeScript support with strict mode', color: 'text-yellow-600' },
-    { icon: '🧹', title: 'Clean Code', description: 'ESLint + Prettier for code quality', color: 'text-gray-600' },
-    { icon: '🎨', title: 'Beautiful UI', description: 'Tailwind CSS with utility-first approach', color: 'text-red-600' }
-  ];
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="container mx-auto px-4 py-16">
-        <div className="text-center mb-16">
-          <h1 className="text-5xl font-bold text-gray-900 mb-6">
-            Welcome to <span className="text-blue-600">PolyFront</span>
-          </h1>
-          <p className="text-xl text-gray-600 mb-6 max-w-3xl mx-auto">
-            Your comprehensive multi-frontend scaffold with everything you need
-          </p>
-          <p className="text-gray-500 max-w-2xl mx-auto">
-            Built with React, TypeScript, Vite, Tailwind CSS, and modern development practices.
-            Start building your next great application today.
-          </p>
-        </div>
-
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 mb-16">
-          {features.map((feature, index) => (
-            <div key={index} className="bg-white rounded-lg shadow-lg p-6 hover:shadow-xl transition-shadow duration-300">
-              <div className="text-4xl mb-4">{feature.icon}</div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">{feature.title}</h3>
-              <p className="text-gray-600">{feature.description}</p>
-            </div>
-          ))}
-        </div>
-
-        <div className="bg-white rounded-lg shadow-lg p-12 text-center">
-          <h2 className="text-3xl font-bold text-gray-900 mb-6">Ready to Get Started?</h2>
-          <p className="text-gray-600 mb-8 max-w-2xl mx-auto">
-            This scaffold includes routing, state management, API integration, testing, and much more.
-            Explore the different pages to see all the features in action.
-          </p>
-          <div className="space-x-4">
-            <button className="bg-blue-600 hover:bg-blue-700 text-white font-semibold py-3 px-6 rounded-lg transition-colors duration-200">
-              View Documentation
-            </button>
-            <button className="border-2 border-blue-600 text-blue-600 hover:bg-blue-600 hover:text-white font-semibold py-3 px-6 rounded-lg transition-colors duration-200">
-              Learn More
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-  }
-
-  // Create other Tailwind pages
-  createOtherTailwindPages(pagesDir, extx, ext, useTS);
-}
-
-function createOtherTailwindPages(pagesDir, extx, ext, useTS) {
-  // About page
-  const about = path.join(pagesDir, `About.${extx}`);
-  if (!fs.existsSync(about)) {
-    writeFileSafe(about, `export default function About() {
-  const techStack = [
-    { name: 'React 19', version: 'Latest', category: 'Framework' },
-    { name: 'TypeScript', version: '5.x', category: 'Language' },
-    { name: 'Vite', version: '7.x', category: 'Build Tool' },
-    { name: 'Tailwind CSS', version: '3.x', category: 'UI Library' },
-    { name: 'React Router', version: '7.x', category: 'Routing' },
-    { name: 'Zustand', version: '5.x', category: 'State Management' },
-    { name: 'Axios', version: '1.x', category: 'HTTP Client' },
-    { name: 'Vitest', version: '3.x', category: 'Testing' }
-  ];
-  const features = [
-    'Modern React with hooks and functional components',
-    'Full TypeScript support with strict mode',
-    'Beautiful Tailwind CSS with utility-first approach',
-    'Client-side routing with React Router',
-    'State management with Zustand',
-    'API integration with Axios',
-    'Comprehensive testing setup with Vitest',
-    'ESLint and Prettier for code quality',
-    'Hot module replacement for fast development',
-    'Responsive design with mobile-first approach'
-  ];
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="container mx-auto px-4 py-16">
-        <div className="text-center mb-16">
-          <h1 className="text-5xl font-bold text-gray-900 mb-6">
-            About <span className="text-blue-600">PolyFront</span>
-          </h1>
-          <p className="text-xl text-gray-600 mb-6 max-w-3xl mx-auto">
-            A comprehensive scaffold for modern web applications
-          </p>
-        </div>
-        <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 mb-16">
-          <div className="bg-white rounded-lg shadow-lg p-8">
-            <h2 className="text-3xl font-bold text-blue-600 mb-4">Technology Stack</h2>
-            <p className="text-gray-600 mb-6">
-              Built with the latest and greatest technologies in the React ecosystem.
-            </p>
-            <div className="grid grid-cols-2 gap-3">
-              {techStack.map((tech, index) => (
-                <div key={index} className="bg-blue-100 text-blue-800 px-3 py-2 rounded-lg">
-                  <div className="font-semibold">{tech.name} {tech.version}</div>
-                  <div className="text-sm text-blue-600">{tech.category}</div>
-                </div>
-              ))}
-            </div>
-          </div>
-          <div className="bg-white rounded-lg shadow-lg p-8">
-            <h2 className="text-3xl font-bold text-blue-600 mb-4">Key Features</h2>
-            <ul className="space-y-3">
-              {features.map((feature, index) => (
-                <li key={index} className="flex items-center">
-                  <div className="w-6 h-6 bg-green-100 rounded-full flex items-center justify-center mr-3">
-                    <div className="w-2 h-2 bg-green-600 rounded-full"></div>
-                  </div>
-                  {feature}
-                </li>
-              ))}
-            </ul>
-          </div>
-        </div>
-        <div className="bg-white rounded-lg shadow-lg p-12 text-center">
-          <h2 className="text-3xl font-bold text-gray-900 mb-8">Why Choose PolyFront?</h2>
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-            <div>
-              <div className="text-4xl mb-4">⭐</div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">Production Ready</h3>
-              <p className="text-gray-600">Built with best practices and industry standards</p>
-            </div>
-            <div>
-              <div className="text-4xl mb-4">⚡</div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">Fast Development</h3>
-              <p className="text-gray-600">Hot reload and instant feedback for rapid iteration</p>
-            </div>
-            <div>
-              <div className="text-4xl mb-4">👥</div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">Team Friendly</h3>
-              <p className="text-gray-600">Consistent code style and comprehensive tooling</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-    // Dashboard page
-    const dashboard = path.join(pagesDir, `Dashboard.${extx}`);
-    if (!fs.existsSync(dashboard)) {
-      writeFileSafe(dashboard, `import { useState } from 'react';
-export default function Dashboard() {
-  const [stats] = useState({
-    users: 1234,
-    revenue: 45678,
-    orders: 89,
-    views: 12345
-  });
-  const [recentActivity] = useState([
-    { id: 1, user: 'Nirmal Samaranayaka', action: 'Order placed', time: '2 min ago', status: 'success' },
-    { id: 2, user: 'Jane Smith', action: 'Payment received', time: '5 min ago', status: 'success' },
-    { id: 3, user: 'Bob Johnson', action: 'Login attempt', time: '10 min ago', status: 'warning' },
-    { id: 4, user: 'Alice Brown', action: 'Account created', time: '15 min ago', status: 'success' },
-  ]);
-  const [topProducts] = useState([
-    { name: 'Premium Widget', sales: 156, revenue: 2340, growth: 12 },
-    { name: 'Super Gadget', sales: 98, revenue: 1890, growth: -5 },
-    { name: 'Mega Tool', sales: 87, revenue: 1450, growth: 8 },
-    { name: 'Ultra Device', sales: 76, revenue: 1230, growth: 15 },
-  ]);
-  const getStatusIcon = (status ${useTS?': string':''}) => {
-    switch (status) {
-      case 'success': return <div className="w-3 h-3 bg-green-500 rounded-full"></div>;
-      case 'warning': return <div className="w-3 h-3 bg-yellow-500 rounded-full"></div>;
-      case 'error': return <div className="w-3 h-3 bg-red-500 rounded-full"></div>;
-      default: return <div className="w-3 h-3 bg-green-500 rounded-full"></div>;
-    }
-  };
-  const getGrowthColor = (growth ${useTS?': number':''}) => growth >= 0 ? 'text-green-600' : 'text-red-600';
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="container mx-auto px-4 py-8">
-        <h1 className="text-4xl font-bold text-gray-900 mb-4">Dashboard</h1>
-        <p className="text-lg text-gray-600 mb-8">
-          Welcome back! Here's what's happening with your application today.
-        </p>
-        {/* Stats Cards */}
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6 mb-8">
-          <div className="bg-white rounded-lg shadow-lg p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600 mb-2">Total Users</p>
-                <p className="text-3xl font-bold text-gray-900 mb-2">{stats.users.toLocaleString()}</p>
-                <p className="text-sm text-green-600 flex items-center">
-                  <span className="mr-1">↗</span>
-                  +12% from last month
-                </p>
-              </div>
-              <div className="w-14 h-14 bg-blue-500 rounded-full flex items-center justify-center">
-                <span className="text-white text-xl">👥</span>
-              </div>
-            </div>
-          </div>
-          <div className="bg-white rounded-lg shadow-lg p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600 mb-2">Revenue</p>
-                <p className="text-3xl font-bold text-gray-900 mb-2">\${stats.revenue.toLocaleString()}</p>
-                <p className="text-sm text-green-600 flex items-center">
-                  <span className="mr-1">↗</span>
-                  +8% from last month
-                </p>
-              </div>
-              <div className="w-14 h-14 bg-green-500 rounded-full flex items-center justify-center">
-                <span className="text-white text-xl">💰</span>
-              </div>
-            </div>
-          </div>
-          <div className="bg-white rounded-lg shadow-lg p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600 mb-2">Orders</p>
-                <p className="text-3xl font-bold text-gray-900 mb-2">{stats.orders}</p>
-                <p className="text-sm text-green-600 flex items-center">
-                  <span className="mr-1">↗</span>
-                  +15% from last month
-                </p>
-              </div>
-              <div className="w-14 h-14 bg-yellow-500 rounded-full flex items-center justify-center">
-                <span className="text-white text-xl">🛒</span>
-              </div>
-            </div>
-          </div>
-          <div className="bg-white rounded-lg shadow-lg p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600 mb-2">Page Views</p>
-                <p className="text-3xl font-bold text-gray-900 mb-2">{stats.views.toLocaleString()}</p>
-                <p className="text-sm text-green-600 flex items-center">
-                  <span className="mr-1">↗</span>
-                  +22% from last month
-                </p>
-              </div>
-              <div className="w-14 h-14 bg-purple-500 rounded-full flex items-center justify-center">
-                <span className="text-white text-xl">👁️</span>
-              </div>
-            </div>
-          </div>
-        </div>
-        {/* Charts and Tables */}
-        <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
-          <div className="lg:col-span-2">
-            <div className="bg-white rounded-lg shadow-lg">
-              <div className="p-6 border-b border-gray-200">
-                <div className="flex justify-between items-center">
-                  <h2 className="text-xl font-semibold text-gray-900">Top Products</h2>
-                  <button className="bg-blue-600 hover:bg-blue-700 text-white px-4 py-2 rounded-lg transition-colors">
-                    View All
-                  </button>
-                </div>
-              </div>
-              <div className="p-6">
-                <div className="overflow-x-auto">
-                  <table className="w-full">
-                    <thead>
-                      <tr className="border-b border-gray-200">
-                        <th className="text-left py-3 px-4 font-medium text-gray-900">Product</th>
-                        <th className="text-right py-3 px-4 font-medium text-gray-900">Sales</th>
-                        <th className="text-right py-3 px-4 font-medium text-gray-900">Revenue</th>
-                        <th className="text-right py-3 px-4 font-medium text-gray-900">Growth</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {topProducts.map((product) => (
-                        <tr key={product.name} className="border-b border-gray-100">
-                          <td className="py-3 px-4">{product.name}</td>
-                          <td className="py-3 px-4 text-right">{product.sales}</td>
-                          <td className="py-3 px-4 text-right">\${product.revenue.toLocaleString()}</td>
-                          <td className="py-3 px-4 text-right">
-                            <span className={\`px-2 py-1 rounded-full text-xs font-medium \${getGrowthColor(product.growth)}\`}>
-                              {product.growth >= 0 ? '+' : ''}{product.growth}%
-                            </span>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="lg:col-span-1">
-            <div className="bg-white rounded-lg shadow-lg">
-              <div className="p-6 border-b border-gray-200">
-                <h2 className="text-xl font-semibold text-gray-900">Recent Activity</h2>
-              </div>
-              <div className="p-6">
-                <div className="space-y-4">
-                  {recentActivity.map((activity) => (
-                    <div key={activity.id} className="flex items-start space-x-3">
-                      <div className="mt-1">
-                        {getStatusIcon(activity.status)}
-                      </div>
-                      <div className="flex-1">
-                        <p className="text-sm font-medium text-gray-900">{activity.action}</p>
-                        <p className="text-xs text-gray-500">
-                          {activity.user} — {activity.time}
-                        </p>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-    // Profile page
-    const profile = path.join(pagesDir, `Profile.${extx}`);
-    if (!fs.existsSync(profile)) {
-      writeFileSafe(profile, `import { useState } from 'react';
-export default function Profile() {
-  const [isEditing, setIsEditing] = useState(false);
-  const [profile] = useState({
-    name: 'Nirmal Samaranayaka',
-    email: 'nirmal.fullstack@gmail.com',
-    phone: '+46 (72) xxx-xxxx',
-    location: 'Stockholm, Sweden',
-    company: 'Scania AB.',
-    position: 'Senior Fullstack Developer',
-    education: 'Computer Science, University of Colombo',
-    website: 'https://dev.to/nirmalsamaranayaka',
-    bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.'
-  });
-  const [settings, setSettings] = useState({
-    emailNotifications: true,
-    pushNotifications: false,
-    darkMode: false,
-    twoFactorAuth: true
-  });
-  const [skills] = useState([
-    'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends', 
-    'TypeScript', 'JavaScript', 'HTML/CSS', 'jQuery', 'Tailwind CSS', 'Bootstrap',
-    '.NET 6/7/8', 'ASP.NET Core', 'C#', 'Web API', 'Microservices', 'Azure', 'AWS'
-  ]);
-  const handleSave = () => {
-    setIsEditing(false);
-  };
-  const handleCancel = () => {
-    setIsEditing(false);
-  };
-  const handleSettingChange = (setting ${useTS?': string':''}) => (event ${useTS?': React.ChangeEvent<HTMLInputElement>':''}) => {
-    setSettings(prev => ({
-      ...prev,
-      [setting]: event.target.checked
-    }));
-  };
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="container mx-auto px-4 py-8">
-        <div className="flex justify-between items-center mb-8">
-          <h1 className="text-4xl font-bold text-gray-900">Profile</h1>
-          <div className="flex space-x-3">
-            {isEditing ? (
-              <>
-                <button 
-                  className="bg-blue-600 hover:bg-blue-700 text-white px-6 py-3 rounded-lg transition-colors"
-                  onClick={handleSave}
-                >
-                  Save
-                </button>
-                <button 
-                  className="bg-gray-300 hover:bg-gray-400 text-gray-700 px-6 py-3 rounded-lg transition-colors"
-                  onClick={handleCancel}
-                >
-                  Cancel
-                </button>
-              </>
-            ) : (
-              <button 
-                className="bg-blue-600 hover:bg-blue-700 text-white px-6 py-3 rounded-lg transition-colors"
-                onClick={() => setIsEditing(true)}
-              >
-                Edit Profile
-              </button>
-            )}
-          </div>
-        </div>
-        <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-          {/* Profile Header */}
-          <div className="lg:col-span-3">
-            <div className="bg-white rounded-lg shadow-lg p-8">
-              <div className="flex items-center space-x-8">
-                <div className="w-32 h-32 bg-blue-600 text-white rounded-full flex items-center justify-center text-4xl font-bold">
-                  {profile.name.split(' ').map(n => n[0]).join('')}
-                </div>
-                <div className="flex-1">
-                  <h2 className="text-3xl font-bold text-gray-900 mb-3">{profile.name}</h2>
-                  <h4 className="text-xl text-gray-600 mb-4">{profile.position} at {profile.company}</h4>
-                  <p className="text-gray-600 mb-6">{profile.bio}</p>
-                  <div className="flex flex-wrap gap-2">
-                    {skills.map((skill) => (
-                      <span key={skill} className="bg-gray-100 text-gray-700 px-3 py-1 rounded-full text-sm">
-                        {skill}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          {/* Profile Details */}
-          <div className="lg:col-span-2">
-            <div className="bg-white rounded-lg shadow-lg">
-              <div className="p-6 border-b border-gray-200">
-                <h3 className="text-xl font-semibold text-gray-900">Personal Information</h3>
-              </div>
-              <div className="p-6">
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Full Name</label>
-                    <input
-                      type="text"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.name}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Email</label>
-                    <input
-                      type="email"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.email}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Phone</label>
-                    <input
-                      type="tel"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.phone}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div>
-                    <label className="text-sm font-medium text-gray-700 mb-2">Location</label>
-                    <input
-                      type="text"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.location}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Company</label>
-                    <input
-                      type="text"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.company}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Position</label>
-                    <input
-                      type="text"
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      value={profile.position}
-                      disabled={!isEditing}
-                    />
-                  </div>
-                  <div className="md:col-span-2">
-                    <label className="block text-sm font-medium text-gray-700 mb-2">Bio</label>
-                    <textarea
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                      rows={3}
-                      value={profile.bio}
-                      disabled={!isEditing}
-                    ></textarea>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          {/* Settings */}
-          <div className="lg:col-span-1">
-            <div className="bg-white rounded-lg shadow-lg">
-              <div className="p-6 border-b border-gray-200">
-                <h3 className="text-xl font-semibold text-gray-900">Settings</h3>
-              </div>
-              <div className="p-6">
-                <div className="space-y-4">
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium text-gray-700">Email Notifications</span>
-                    <input
-                      type="checkbox"
-                      className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
-                      checked={settings.emailNotifications}
-                      onChange={handleSettingChange('emailNotifications')}
-                    />
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium text-gray-700">Push Notifications</span>
-                    <input
-                      type="checkbox"
-                      className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
-                      checked={settings.pushNotifications}
-                      onChange={handleSettingChange('pushNotifications')}
-                    />
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium text-gray-700">Dark Mode</span>
-                    <input
-                      type="checkbox"
-                      className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
-                      checked={settings.darkMode}
-                      onChange={handleSettingChange('darkMode')}
-                    />
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium text-gray-700">Two-Factor Auth</span>
-                    <input
-                      type="checkbox"
-                      className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 focus:ring-2"
-                      checked={settings.twoFactorAuth}
-                      onChange={handleSettingChange('twoFactorAuth')}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className="bg-white rounded-lg shadow-lg mt-6">
-              <div className="p-6 border-b border-gray-200">
-                <h3 className="text-xl font-semibold text-gray-900">Quick Actions</h3>
-              </div>
-              <div className="p-6">
-                <div className="space-y-3">
-                  <button className="w-full bg-gray-100 hover:bg-gray-200 text-gray-700 px-4 py-2 rounded-lg transition-colors">
-                    Download Resume
-                  </button>
-                  <button className="w-full bg-gray-100 hover:bg-gray-200 text-gray-700 px-4 py-2 rounded-lg transition-colors">
-                    Share Profile
-                  </button>
-                  <button className="w-full bg-red-100 hover:bg-red-200 text-red-700 px-4 py-2 rounded-lg transition-colors">
-                    Delete Account
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
- );
-}`);
-    }
-  }
-}
 }
 
 function createAntdPages(pagesDir, extx, ext, useTS) {
@@ -2973,7 +2400,7 @@ const Layout = () => {
           </Typography>
 
           <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-            <Chip label="v0.0.27" size="small" variant="outlined" />
+            <Chip label="v0.0.28" size="small" variant="outlined" />
             {/* Language selector using i18n */}
             <Select
               value={currentLang}
@@ -3118,7 +2545,7 @@ const Layout = () => {
             </button>
             <h4 className="mb-0">PolyFront</h4>
             <div className="d-flex align-items-center gap-2">
-              <span className="badge bg-secondary">v0.0.27</span>
+              <span className="badge bg-secondary">v0.0.28</span>
               
               {/* Language Selector */}
               <select
@@ -3212,7 +2639,7 @@ export default Layout;`);
                 {/* Version + Language + User */}
                 <div className="flex items-center gap-3">
                   {/* Version */}
-                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v0.0.27</span>
+                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v0.0.28</span>
       
                   {/* Language selector */}
                   <select
@@ -3321,7 +2748,7 @@ style={{ borderRight: 0 }}
         >
           PolyFront
         </Typography.Title>
-        <Tag bordered={false}>v0.0.27</Tag>
+        <Tag bordered={false}>v0.0.28</Tag>
         <Select
           size="small"
           value={currentLang}
@@ -3515,7 +2942,7 @@ function Layout() {
           <Box flex="1" />
 
           <HStack gap={2}>
-            <Badge>v0.0.27</Badge>
+            <Badge>v0.0.28</Badge>
 
             {/* Language Select (v3 controlled) */}
             <Select.Root
